requirements.txt
botocore==1.15.32
pyodbc==4.0.28
boto3==1.12.32
xmltodict==0.12.0
requests==2.22.0
azure-log-analytics-data-collector-api==0.2.0
